Fixes bugs in the addition and indexing of a string.
One of the bugs was in the addition of two string. When the values were
strings, then the "are the types equal?" condition was met, but that one
doesn't handle strings. Now it's fixed.

The second one: I was confused about char to string conversion. Justifiably.
(C++... are you drunk?)