Fixes the issue with garbage collecting lists.
This is the problem we had with hangman.odo.
The fix was hard to find, but simple in nature. I was copying the list when passing it through functions, which shouldn't happen.

In the garbage collecting section, I only need to remove the references to the list_element symbols when I'm sure I can destruct the list.