Adds basic OOP to Odo.
I'm not too sure about inheritance yet, but at least the basic functions work now.

On the topic of inheritance, getting that to work had a high cost. I'll try trade my soul back in a future commit.

In each instance there's a linked list of SymbolTables, but I create the middle ones, and they need to be pointers, so...

Raw pointers that I'm not disposing later...

Ugh.