Fixes a bug where null was inconsistently corrupted.
It feels so good. Too little lines. 2 or 3 days have I been fighting this bug. Non of the tests I had before checked for this bug, so all of them worked fine when I made the type of a value a pointer.

The problem was that for null I was assigning it a pointer to the NullType symbol in the temporary map in the interpreter constructor.

Now null gets its type from the permanent symbol in the globalTable.

Nice.
<3