Fixes bug in forange statement where its iterator was discarded prematurely.
The iterator in the FoRange gets a new value each time, a value created at the moment. This means the GB considers it discardable.

I could attach the reference to the value, but that is really expensive and I don't really see the need yet.
For the moment, it's just marked as important.

Fixed.

<3