Optimizes previous fix on attaching a reference to value in for* loop.
I went and bloody did it.
Now there is just one symbol with one value, instead of creating a new value each iteration.

Since the forange loop iterates while just changing the number it represents, setting the underlying val and keeping the Value instance gets rid of unnecessary bindings.

The foreach loop is the same, but just in the case of iterating over a string.
When using it on a list it is assured that the value has a symbol attached to it.

I'm proud of this one.
<3