Possible language design idea.
How about that?
The same concept I made for foreach and forange, but this ones generate a list and return the value.

The concept is similar to a list comprehension in python or like Array.map in javascript.

I'll think about it, because this could be making the language a little bloat.