Adds a special variable to the repl that show the last result.
There shouldn't be much problem with this change. Just a quality of life improvement.
I copied it because Andreas Kling showed the latest improvements to serenityOS, and someone added this exact feature to the javascript Interpreter.

For the record, the primitive base for odo, the arduino calculator, had this too, so it's not completely copied (like the rest of the language).

This commit message is bigger that the actual change.

:^)