Refactors value to support inheritance. (unfinished)
Previously the structure value stored every possible content that a value could have, but that wasn't as efficient in memory and performance.

The code was ugly too.

Next is refactoring the Symbol module.
(and I guess AST later?)

Then I want to make different files for each class in the modules.

<3