Adapts the parser to use the new ast node system.
It's really nice seeing the parser create an instance of every class I actually need, with exactly the elements it needs to function, instead of carrying around every appendix variable that an AST _could_ have.

While refactoring this I noticed how many times I would be setting the line number on a node, but that isn't that important.