Updates Interpreter with the new Value and Node system (compiles)
So cool!
Yes!!!

The interpreter can now use polymorphism in the node tree to check which type a node is, cast it, and visit it without unfolding it like before.

With smart pointers I can also relax a little about the lifetime of my objects.

Creating and maintaining values doesn't change that much, but with inheritance I can keep every object containing only exactly what it needs. This also lets me think about making a standard library, and putting the creation of the bindings from odo to native code in separate modules.

I'll have to keep testing this to find bugs, though.

Cool.
Cool, cool, cool.