Gets rid of the ValueTable (optimization)
The value table was an appendix from the times I used swift to write this program. I needed it to keep track of values. It simulated a heap (I think? Computer architecture is not my forte). It kept values stored in a safe place, so everyone else could point to that memory. It also kept reference count to be able to remove unused pointers.

Either way, the valueTable was useful for me. But.
Now that I made the values be stored in the actual heap with smart reference counting pointers, the value table has lost much of it's purpose.

This was always the slowest part of the runtime, so this is a really good improvement.