Removes the reference counting inside of value.
This was only needed as part of the valueTable garbage collection. Now that the smart pointers keep track of all of that, I don't really need the reference counting.

Maybe I will implement a way to reference the first symbol that a value gets, just so I can get a default name for all values.