Changes the exception api to return messages as string objects instead of const char*
In macOS this didn't seem to be an actual problem, but in Windows the object that gave its address gets destroyed before displaying, causing weird bugs on the console.

It also changes the name of the function that returns said message, which deviates from the standard, but it's hardly a problem. And I would prefer function over tradition.