Adds more support for spanish flavored odo.
This is the rest of the translations. Nothing much to add. I'm quite excited for this, since it makes the translation effort so much more streamlined.

Also, this means that I now have every string in the language in a single place, and every place that could need change will receive it at once.