Solves bug in the pop function.
This a bit of a hack, but apparently the value that was popped from the list gets destroyed because there are no more references to it.

I solved it by temporary storing the value in the 'returning' member variable.

I also clear the variable returning on every call if it isn't null.