Adds support for visit_ListDeclaration to SemAn.
This is a bigger scale addition, and show a little more about the SemAn.

The fact that I can check what type the list has and decide if the initializer is right feels so nice.

I'm feeling uneasy about the error messages, though. I think I need to work a little more on them.