Refactors the behavior of the function context and function definition in SemAn.
I was having problems with the way I set up the function declaration before.

The fact that it was defined until after I was sure there were no semantic problems meant that it's  function body couldn't find it or use it recursively.

The solution was having the definition before the body, and if the body threw an error, removing the definition from the scope.

I also had to change the way I handle function's semantic context.

I didn't want it's parent scope to be the same as the first time that function type was defined, so instead I make a copy of the context, change the parent to the current scope, and use it like that.