Refactors vardecl in case the type is defined in the initialization.
Like with function expressions!

There's a possibility that when the function is generated, the type of the function will also be created, so I needed to get the visit(node->initial) above the request for the new variable's type.

This is also a secondary effect of the SemAn giving a declaration with "any" as its type, the type of the initialization.

A lot o' words, man.

Damn.