Changes the syntax for function expressions body.
Now I can ignore the arrow when defining the function's body as just a return statement.

This could be the new thing inn Odo.

Instead of:
`if condition { one_statement }`
I could do:
`if condition -> one_statement`.

And so on with every thing that takes just one statement as body.

For some reason the shortest changes make for the longest commit messages.