Adds support for visit_FuncExpression to SemAn.
Big addition.
This handles it's own semantic context and type definition, sets up the type inference that funcExpression relies on...

For some reason, in interpreter I was assuming that this functions could have a type defined.
But the parser never allowed that.

But now it's possible because the SemAn can set it before.