Adds support for member variables in the SemAn, with inheritance.
This is a little messy.
I had to create the raw pointers for the inheritance scopes in instances again, but this time I had a plan for deleting them later.

I general, this feels like an improvement over the language, but I discovered a problem arising.

The semantic analyzer means that everything will be sequential, and one function will not be able to call another before it's defined.

Like:
```
func main() {
    hello()
}

func hello() {
    println("Hello, world!")
}
```

Would be an error, because the SemAn doesn't know that hello exists yet.

I'll think of a few solutions.

Maybe a cache of symbols that only gets checked once sed for the first time.