Adds error handling when creating a module. If it fails, undefine it.
Just to be sure.

When testing this after adding the try catch, I was so happy to see that I could just fix the error in the file and re import the module on the spot.

This language is really feeling like it's strongly typed, but the interpreted aspect of it is so nice.

Cool. Cool, cool, cool.