Fixes bug in SemAn where an empty list couldn't be assigned to any list because of type incompatibility.
I just made it so an empty list automatically took the type of the variable it's being assigned to.

Fixes bug in SemAn where an empty list couldn't be assigned to any list because of type incompatibility.

I just made it so an empty list automatically took the type of the variable it's being assigned to.