Bug Fix: You could try to assign to a function type alias.
This caused undefined behavior, because the function type could or could not have a return type, which is what assignment uses to check type compatibility.

I added a check for the type that the symbol to assign has, and also a check to make sure that you're not reassigning to a type.