Refactor: Adds full support for list types.
This makes it possible to get and generate the fully formed list type when knowing, or figuring out from the type name, what the dimension of the list is.

This should make it possible to create function aliases for any type, and it also allows the SemAn to infer the type of list elements, no matter how deep is the list.

Cool. Cool, cool, cool.