Refactor: Changes legacy bool operators && and || to theh actual words.
At first I wanted to have them because i was trying to keep the language as C-like as possible but without the old parts. "and", "or" felt really python-y.

But at this point, I'm obviously not gonna implement bitwise operators.
At least not with those exact characters.

And the symbols don't really mean nothing anymore, and that's the kind of legacy artifcats I'm trying to remove from odo.

Besides, if this language is gonna be used for teaching to learn programming, in spanish, then it's gonna be much easier to explain what "and" means than "&&".

So yeah.
Cool stuff.