NewFtr: Adds support for NativeFunctions in NativeModule.
I'm resisting the temptation of writing a short commit message. 9 files changed, and I want to leave this at the title.

This commit has some dubious code, and the API can be improved. but overall, I'm proud of the result and excited about the next steps.

I'm really excited about this. It makes Odo a little more formal and organized.

The Semantic Analyzer is proving to be useful right now, as I only have to setup the param and return type for the functions, and the interpreter can simply visit parameters and return the function's value, trusting that everything has the right type.

Reaaaaally cool stuff.