BugFix: Now a class creates a lazy scope frame so that the instance exists as soon as the class is checked.
I think this is the best of both worlds. An instance is just an extension of the class' body, so it doesn't need to wait for longer than the class.

THe fact that it waits til the scope is consumed assures that the static components of the class exist by the time the instance is checked.

Again, 2 lines of code with an essay of a commit message.