NewFtr: Implements simple command line argument handling.
This commit also handles catching any kind of program exit so that I can clean up.

I need to be able to reset the terminal colors, specially when ctrl-c in the repl.
Else, everything will be yellow afterwards.

The library flags seems to be so simple.
Thanks to the developers, and thanks for making it public domain.