NewFtr: Adds basic test suite for testing implementation.
This is not that thorough, but it will help me keep track of what I break and what still works on every change to odo.

The tests run after every release build, so I can check quickly that everything still works.

The problem is... I need to write a test for everything I need to protect from first.